+++ Android/UdacityExercises/Lesson10-Hydration-Reminder/T10.04-Exercise-PeriodicSyncWithJobDispatcher/app/src/main/java/com/example/android/background/sync/WaterReminderFirebaseJobService.java
@@ -48,7 +48,7 @@ public class WaterReminderFirebaseJobService extends JobService {
     @Override
     public boolean onStopJob(JobParameters jobParameters) {
         if (mBackgroundTask != null)
-            mBackgroundTask.cancel(true)
+            mBackgroundTask.cancel(true);
 
         return true;
     }