+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 22.09.2018
 
+* [Android] Hydration Reminder: Schedule with FirebaseJobDispatcher 
 * [Android] Hydration Reminder: Adding a JobService
 
 #### 21.09.2018