+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 23.09.2018
+
+* [Android] Hydration Reminder: Firebase experiments
+
 #### 22.09.2018
 
 * [Android] Hydration Reminder: Schedule with FirebaseJobDispatcher